backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dotenv==1.0.1
email-validator==2.2.0

celery==5.4.0
redis==5.2.1

PyJWT==2.9.0
passlib[bcrypt]==1.7.4

google-generativeai==0.8.4
